Move decoder settings to its own module
